Initial cleanup and dependency junit
