+++ pom.xml
@@ -16,21 +16,20 @@
 			<name>Vidal Fontoura</name>
 			<email>vidalfontoura16@gmail.com</email>
 		</developer>
+		<developer>
+			<name>Ricardo Lima</name>
+			<email>ricardo_des@hotmail.com</email>
+		</developer>
 	</developers>
 
-
-
-
-	
-	
-
 	<dependencies>
-	<dependency>
-	<groupId>commons-lang</groupId>
-	<artifactId>commons-lang</artifactId>
-	<version>2.2</version>
-</dependency>
 	
+		<dependency>
+			<groupId>junit</groupId>
+			<artifactId>junit</artifactId>
+			<version>4.12</version>
+		</dependency>
+		
 	
 	</dependencies>
