--- src/main/java/org/ufpr/cbio/poc/Grid.java
@@ -1,10 +1,0 @@
-package org.ufpr.cbio.poc;
-
-/**
- *
- *
- * @author user
- */
-public class Grid {
-
-}
