+++ src/main/java/org/ufpr/cbio/poc/main/Main.java
@@ -3,13 +3,11 @@
  */
 package org.ufpr.cbio.poc.main;
 
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
 import org.ufpr.cbio.poc.domain.Grid;
 import org.ufpr.cbio.poc.domain.Residue;
-import org.ufpr.cbio.poc.domain.ResidueType;
 import org.ufpr.cbio.poc.domain.TopologyContact;
 import org.ufpr.cbio.poc.utils.ResidueUtils;
 
@@ -20,9 +18,9 @@ import org.ufpr.cbio.poc.utils.ResidueUtils;
  */
 public class Main {
 
-    // private static final String CHAIN_SEQUENCE = "HHHPHPPPPPH";
+    private static final String CHAIN_SEQUENCE = "HHHPHPPPPPH";
 
-    private static final String CHAIN_SEQUENCE = "PPHPHHHHHHP";
+    // private static final String CHAIN_SEQUENCE = "PPHPHHHHHHP";
 
     public static void main(String[] args) {
 
@@ -34,46 +32,10 @@ public class Main {
 
         Grid grid = new Grid(CHAIN_SEQUENCE.length(), CHAIN_SEQUENCE.length());
 
-        int[][] matrix = grid.buildResidueStructure(residues);
-        Set<TopologyContact> topologyContacts = new HashSet<>();
+        grid.buildResidueStructure(residues);
+        grid.printResidueStructure();
 
-        int index = 0;
-        for (int i = 0; i < residues.size(); i++) {
-            if (residues.get(i).getResidueType().equals(ResidueType.P)) {
-                continue;
-            }
-            if (residues.get(i).getPoint().y + 1 < matrix.length) {
-                index = matrix[residues.get(i).getPoint().y + 1][residues.get(i).getPoint().x];
-                // test up
-                if (isTopologicalContact(i, index, residues)) {
-                    topologyContacts.add(new TopologyContact(residues.get(i), residues.get(index)));
-                }
-            }
-            if (residues.get(i).getPoint().x + 1 < matrix.length) {
-                // test right
-                index = matrix[residues.get(i).getPoint().y][residues.get(i).getPoint().x + 1];
-                if (isTopologicalContact(i, index, residues)) {
-                    topologyContacts.add(new TopologyContact(residues.get(i), residues.get(index)));
-                }
-            }
-            if (residues.get(i).getPoint().y - 1 >= 0) {
-                // test down
-                index = matrix[residues.get(i).getPoint().y - 1][residues.get(i).getPoint().x];
-                if (isTopologicalContact(i, index, residues)) {
-                    topologyContacts.add(new TopologyContact(residues.get(i), residues.get(index)));
-                }
-            }
-            if (residues.get(i).getPoint().x - 1 >= 0) {
-                // test back
-                index = matrix[residues.get(i).getPoint().y][residues.get(i).getPoint().x - 1];
-                if (isTopologicalContact(i, index, residues)) {
-                    topologyContacts.add(new TopologyContact(residues.get(i), residues.get(index)));
-                }
-            }
-
-        }
-
-        // topologyContacts = removeDuplicates(topologyContacts);
+        Set<TopologyContact> topologyContacts = ResidueUtils.getTopologyContacts(residues, grid);
         System.out.println(topologyContacts.size());
         for (TopologyContact topologyContact : topologyContacts) {
             System.out.print("r1[" + topologyContact.getR1().getPoint().getX() + ","
@@ -81,29 +43,6 @@ public class Main {
             System.out.println("r2[" + topologyContact.getR2().getPoint().getX() + ","
                 + topologyContact.getR2().getPoint().getY() + "]");
         }
-
     }
 
-    public static boolean isTopologicalContact(int i, int index, List<Residue> residues) {
-
-        if (i != index + 1 && i != index - 1 && index != -1) {
-
-            return residues.get(index).getResidueType().equals(ResidueType.H);
-        }
-        return false;
-    }
-
-    public static Set<TopologyContact> removeDuplicates(Set<TopologyContact> topologyContacts) {
-
-        Set<TopologyContact> contacts = new HashSet<>();
-        for (TopologyContact topologyContact : topologyContacts) {
-            for (TopologyContact topologyContact2 : topologyContacts) {
-                if (topologyContact.getR1().getPoint().getX() == topologyContact2.getR2().getPoint().getX()
-                    && topologyContact.getR1().getPoint().getY() == topologyContact2.getR2().getPoint().getY()) {
-                    contacts.add(topologyContact);
-                }
-            }
-        }
-        return contacts;
-    }
 }
